generate advs on resnet

+++ CNN/GBP_Denoising/attack.py
@@ -190,7 +190,7 @@ def attack_one_image(image, name, label, attack_type, fool_model):
 
                     print('The attack is successed!')
 
-                    simple_plot(adversarial, 'ADV' + name, './Adversarial_Examples/{}/'.format(attack_type))
+                    simple_plot(adversarial.astype(int), 'ADV' + name, './Adversarial_Examples/{}/'.format(attack_type))
 
                     print('Saved!')
